requirements.txt
Django==1.10.4
httpie==0.9.9
gunicorn==19.6.0
Pillow==3.4.2
PyMySQL==0.7.9
amqp==1.4.9
anyjson==0.3.3
arrow==0.8.0
billiard==3.5.0.2
binaryornot==0.4.0
beautifulsoup4==4.5.1
celery==4.0.2
chardet==2.3.0
click==6.6
cookiecutter==1.4.0
Unidecode==0.4.19
django-cors-headers==1.1.0
django-debug-panel==0.8.3
django-debug-toolbar==1.4
django-countries==4.0
django-factory-boy==1.0.0
django-filter==1.0.1
django-jenkins==0.110.0
django-modelcluster==2.0
django-overextends==0.4.2
django-taggit==0.21.3
django-treebeard==4.1.0
djangorestframework==3.5.0
django-dbbackup==3.0.0
html5lib==0.9999999
factory-boy==2.7.0
fake-factory==0.5.9
future==0.15.2
fields==5.0.0
Jinja2==2.8
jinja2-time==0.2.0
kombu==4.0.1
MarkupSafe==0.23
poyo==0.4.0
python-dateutil==2.6.0
pytz==2016.10
requests==2.12.4
setuptools==32.1.2
simplejson==3.10.0
six==1.10.0
sphinx-me==0.3
sqlparse==0.2.2
vine==1.1.3
whichcraft==0.4.0
Willow==0.4
virtualenv==1.9
wagtail==1.8
wheel==0.29.0